+++ src/TmdbServiceProvider.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace BjTheCod3r\Tmdb;
 
 use BjTheCod3r\Tmdb\Client\TmdbClient;
+use BjTheCod3r\Tmdb\Support\ImageUrl;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Support\ServiceProvider;
 
@@ -18,8 +19,14 @@ class TmdbServiceProvider extends ServiceProvider
             return new TmdbClient($app['config']->get('tmdb', []));
         });
 
+        $this->app->singleton(ImageUrl::class, function (Application $app) {
+            return new ImageUrl(
+                $app['config']->get('tmdb.image_base_url', 'https://image.tmdb.org/t/p/'),
+            );
+        });
+
         $this->app->singleton(Tmdb::class, function (Application $app) {
-            return new Tmdb($app->make(TmdbClient::class));
+            return new Tmdb($app->make(TmdbClient::class), $app->make(ImageUrl::class));
         });
 
         // Allow resolving the manager via the "tmdb" container alias.